+++ CHANGELOG.md
@@ -28,6 +28,9 @@ This version is a new scaffolding of cronos project where ethermint is included 
 - (ethermint) [tharsis#383](https://github.com/tharsis/ethermint/pull/383) `GetCommittedState` use the original context.
 
 ### Features
+
+- [#11](https://github.com/crypto-org-chain/cronos/pull/11) embed gravity bridge module
+
 ### Improvements
 
 - (ethermint) [tharsis#425](https://github.com/tharsis/ethermint/pull/425) Support build on linux arm64